requirements.txt
mlflow==2.6.0
boto3
botocore
psycopg2
psycopg2-binary
python-dotenv==0.21.0
